requirements.txt
pytest==5.4.1
selenium==3.141.0
pytest-xdist==1.31.0
allure-pytest==2.8.13